requirements.txt
harness-py-sdk
PyYAML